refactor: move constants and parser to utils
